+++ shared/src/commonMain/kotlin/data/credentials/CredentialAttributeTranslator.kt
@@ -4,6 +4,7 @@ import at.asitplus.jsonpath.core.NormalizedJsonPath
 import at.asitplus.wallet.companyregistration.CompanyRegistrationScheme
 import at.asitplus.wallet.cor.CertificateOfResidenceScheme
 import at.asitplus.wallet.eupid.EuPidScheme
+import at.asitplus.wallet.eupidsdjwt.EuPidSdJwtScheme
 import at.asitplus.wallet.healthid.HealthIdScheme
 import at.asitplus.wallet.idaustria.IdAustriaScheme
 import at.asitplus.wallet.lib.data.ConstantIndex
@@ -19,6 +20,7 @@ interface CredentialAttributeTranslator {
         operator fun get(scheme: ConstantIndex.CredentialScheme?) = when(scheme) {
             is IdAustriaScheme -> IdAustriaCredentialAttributeTranslator
             is EuPidScheme -> EuPidCredentialAttributeTranslator
+            is EuPidSdJwtScheme -> EuPidCredentialAttributeTranslator
             is MobileDrivingLicenceScheme -> MobileDrivingLicenceCredentialAttributeTranslator
             is PowerOfRepresentationScheme -> PowerOfRepresentationCredentialAttributeTranslator
             is CertificateOfResidenceScheme -> CertificateOfResidenceCredentialAttributeTranslator